factor out sample REST response
